:gem: Use IO.popen to access copy command.
I was investigating the need to use `echo -e` to process
backslash-escaped characters on Linux and was having a ton of trouble
getting the `` command to put the emoji character through the pipe in a
way that it be interpreted correctly. I also ran into the "hanging"
issue described in bintest/emoji-cli.rb#14 which an strace revealed was
a mysteriously un-closed file descriptor for the child's standard input.

Since the mruby-io mrbgem was already present, I rewrote the copy method
to use IO.popen rather than backticks to directly write the emoji
character to the standard input of the copy command's process.

I haven't yet been able to test this for Mac OS but we can always make
it a Linux specific change if it doesn't work or the the backtick of echo is
preferred.

I still don't know exactly _why_ the echo command wasn't working on
Linux and I'm a bit too unfamiliar with mruby to really dig into it.